+++ src/pages/Dashboard/UpcomingRuns-Tile.vue
@@ -196,6 +196,11 @@ export default {
           }"
           @click="tab = 'upcoming'"
         >
+          {{
+            upcomingRuns.length > 0 && tab === 'late'
+              ? `(${upcomingRuns.length})`
+              : ''
+          }}
           Upcoming
           <v-icon small>access_time</v-icon>
         </v-btn>
@@ -223,6 +228,11 @@ export default {
           <v-icon v-if="lateRuns && lateRuns.length > 0" small color="deepRed">
             warning
           </v-icon>
+          {{
+            lateRuns.length > 0 && tab === 'upcoming'
+              ? `(${lateRuns.length})`
+              : ''
+          }}
           Late
           <v-icon small>timelapse</v-icon>
         </v-btn>